backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
pydantic-settings
python-dotenv
alembic
openai
anthropic
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
aiosmtplib
# Added during development
httpx
google-generativeai
python-jose
passlib[bcrypt]
python-multipart
greenlet
